changed error message
The problem isn't with the seed being `0`, it's with the argument of `rad` being `0`.  `rad` takes an atom `n` and produces an atom in the range of `0` to `n-1` (inclusive), not including `n`.  But when `n=0` this is a problem!  The error message now reads `rad-underflow` by analogy with `decrement-underflow`